refactor(cosmo_val): extract RealSpaceMixin (real_space)
